module loader template: pass project and version information to loader
This requires that the project version information is generated during
the build process and then this is passed down to the module loader class.

If this information is not available, then simple fallback values are used.

At the end, this allows to call the module on command line with parameter --version.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>